Keep messages delayed for more a long term only on disk
We've seen use cases when messages are delayed for hours, days or even a
month. There is no need to keep all of them in memory. This change
introduces a long-term, disk-only table for these messages. They are
still present in the index table but a field in the key indicates in
which table the actual delivery is stored. It is possible that for the
same timestamp there are both entries in the long-term table scheduled a
long time ago and in the short-term table scheduled recently. Because
?SHORTTERM < ?LONGTERM, the recently scheduled entries preceed the
long-term entries in the ordered_set index table and are delivered
first.

This way not only the runtime memory is reduced but also the load time
of the main, short-term table at startup.